Covert source to ES6 modules
