+++ .eslintrc
@@ -1,3 +1,3 @@
 {
-  "extends": ["frost-standard"]
+  "extends": ['frost-standard']
 }
