+++ .gitignore
@@ -3,6 +3,9 @@
 # dependencies
 /node_modules
 
+# compiled code
+/lib
+
 # misc
 /coverage/*
 /libpeerconnection.log
