+++ package.json
@@ -4,9 +4,10 @@
   "description": "Core bunsen library",
   "main": "lib/index.js",
   "scripts": {
+    "build": "babel src -d lib --presets es2015 --plugins add-module-exports",
     "lint": "eslint *.js lib tests",
     "test": "npm run lint && npm run utest",
-    "utest": "istanbul cover _mocha -- --recursive tests"
+    "utest": "npm run build && istanbul cover _mocha -- --recursive tests"
   },
   "repository": "git@github.com:ciena-blueplanet/bunsen-core.git",
   "engines": {
@@ -22,6 +23,9 @@
   },
   "homepage": "https://github.com/job13er/bunsen-core#readme",
   "devDependencies": {
+    "babel-cli": "^6.9.0",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-preset-es2015": "^6.9.0",
     "chai": "^3.5.0",
     "chai-subset": "^1.2.2",
     "eslint": "^2.11.1",
